update: rename ethashproof package
