+++ core/cache.go
@@ -1,4 +1,4 @@
-package ethashproof
+package core
 
 import (
 	"encoding/json"
