+++ core/dag_merkle_proof.go
@@ -1,4 +1,4 @@
-package ethashproof
+package core
 
 import (
 	"fmt"
