+++ core/utils.go
@@ -1,4 +1,4 @@
-package ethashproof
+package core
 
 import (
 	"github.com/ethereum/go-ethereum/core/types"
